# GestionCandidat/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/cloture/formation/{formation}": {
            "post": {
                "tags": [
                    "Formations"
                ],
                "summary": "Close a formation",
                "description": "Closes a specific formation",
                "operationId": "closeFormation",
                "parameters": [
                    {
                        "name": "formation",
                        "in": "path",
                        "description": "ID of the formation to be closed",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Formation closed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "Formation": {
                                            "type": "string",
                                            "example": "Cloture/en cours"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Formation not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Formation not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/ajout/formation": {
            "post": {
                "tags": [
                    "Formations"
                ],
                "summary": "Add a new formation",
                "description": "Creates a new formation",
                "operationId": "addFormation",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nomFormation",
                                    "description",
                                    "duree",
                                    "status"
                                ],
                                "properties": {
                                    "nomFormation": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Formation XYZ"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Description of the formation"
                                    },
                                    "duree": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "en_cours",
                                            "en_attente"
                                        ],
                                        "example": "en_cours"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Formation created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "Formation": {
                                            "$ref": "#/components/schemas/Formation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "field_name": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "The field_name field is required."
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/modifier/formation/{formation}": {
            "post": {
                "tags": [
                    "Formations"
                ],
                "summary": "Update a formation",
                "description": "Updates a specific formation",
                "operationId": "updateFormation",
                "parameters": [
                    {
                        "name": "formation",
                        "in": "path",
                        "description": "ID of the formation to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nomFormation",
                                    "description",
                                    "duree",
                                    "status"
                                ],
                                "properties": {
                                    "nomFormation": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Updated Formation Name"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated Description of the formation"
                                    },
                                    "duree": {
                                        "type": "integer",
                                        "example": 20
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "en_cours",
                                            "en_attente"
                                        ],
                                        "example": "en_cours"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Formation updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "Formation": {
                                            "$ref": "#/components/schemas/Formation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "field_name": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "The field_name field is required."
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/supprimer/formation/{formation}": {
            "delete": {
                "tags": [
                    "Formations"
                ],
                "summary": "Delete a formation",
                "description": "Deletes a specific formation",
                "operationId": "deleteFormation",
                "parameters": [
                    {
                        "name": "formation",
                        "in": "path",
                        "description": "ID of the formation to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Formation deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "Formation": {
                                            "type": "string",
                                            "example": "Formation supprimée avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Formation not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Formation not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    }
}